# project/docs/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Auditable Agent Runtime API
  version: 1.0.0
  description: Minimum MVP contract for a multi-tenant auditable Agent Runtime.
servers:
  - url: /api/v1
security:
  - bearerAuth: []
paths:
...
  /admin/settings:
    get:
      summary: Get admin settings
      description: Returns global system settings (admin only).
      operationId: getAdminSettings
      responses:
        '200': { description: Settings returned }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
    put:
      summary: Update admin settings
      description: Updates global system settings (admin only).
      operationId: updateAdminSettings
      requestBody:
        required: true
        content: { application/json: { schema: { type: object } } }
      responses:
        '200': { description: Settings updated }
        '400': { $ref: '#/components/responses/ValidationError' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '403': { $ref: '#/components/responses/Forbidden' }
...
  /tenants/{tenant_id}/quota:
    get: { summary: Get tenant quota, description: Returns quota configuration and usage., operationId: getTenantQuota, parameters: [{ $ref: '#/components/parameters/TenantId' }], responses: { '200': { description: Quota returned }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
    put: { summary: Update tenant quota, description: Replaces tenant quota limits., operationId: updateTenantQuota, parameters: [{ $ref: '#/components/parameters/TenantId' }], requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: Quota updated }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' }, '404': { $ref: '#/components/responses/NotFound' }, '500': { description: Internal error } } }
  /settings/profile:
    get: { summary: Get profile, description: Returns the authenticated user's profile and preferences., operationId: getProfile, responses: { '200': { description: Profile returned }, '401': { $ref: '#/components/responses/Unauthorized' } } }
    put: { summary: Update profile, description: Updates the authenticated user's profile and preferences., operationId: updateProfile, requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: Profile updated }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' } } }
  /settings/password:
    put: { summary: Change password, description: Changes the authenticated user's password., operationId: updatePassword, requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: Password updated }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' } } }
  /settings/tenant:
    get: { summary: Get settings quota, description: Returns tenant settings for administrators., operationId: getSettingsTenant, responses: { '200': { description: Tenant settings returned }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } } }
    put: { summary: Update settings quota, description: Updates tenant settings for administrators., operationId: updateSettingsTenant, requestBody: { required: true, content: { application/json: { schema: { type: object } } } }, responses: { '200': { description: Tenant settings updated }, '400': { $ref: '#/components/responses/ValidationError' }, '401': { $ref: '#/components/responses/Unauthorized' }, '403': { $ref: '#/components/responses/Forbidden' } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: opaque }
  parameters:
...
    TenantId: { name: tenant_id, in: path, required: true, schema: { type: string } }
  responses:
    Unauthorized: { description: Authentication required, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    Forbidden: { description: Action is not allowed, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    NotFound: { description: Resource not found, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
    ValidationError: { description: Invalid request, content: { application/json: { schema: { $ref: '#/components/schemas/ErrorEnvelope' } } } }
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message, requestId]
          properties:
            code: { type: string, enum: [UNAUTHORIZED, FORBIDDEN, NOT_FOUND, VALIDATION_ERROR, CONFLICT, VERSION_CONFLICT, RATE_LIMIT_EXCEEDED, QUOTA_EXCEEDED, BUDGET_EXCEEDED, INVALID_STATE_TRANSITION, APPROVAL_ALREADY_HANDLED, RESUME_WINDOW_EXPIRED, MODEL_PROVIDER_ERROR, TOOL_EXECUTION_FAILED, SANDBOX_TIMEOUT, INTERNAL_ERROR] }
            message: { type: string, maxLength: 500 }
            requestId: { type: string }
            details: { type: object, additionalProperties: true }
...